chore: moved language select into component
