+++ frontend/src/routes/Navbar.svelte
@@ -14,25 +14,11 @@
 	import '../app.postcss';
 	import { locale } from 'svelte-i18n';
 	import { browser } from '$app/environment';
+	import LanguageSelect from '$lib/LanguageSelect.svelte';
 
 	export let fluid = true;
 
 	export let drawerHidden: boolean;
-
-	let locales = [
-		{ name: 'English', value: 'en' },
-		{ name: 'German', value: 'de' }
-	];
-
-	let selected = $locale || 'en';
-	console.log(selected);
-	$: {
-		$locale = selected;
-		// also set cookie if browser
-		if (browser) {
-			document.cookie = `locale=${selected};path=/;max-age=31536000`;
-		}
-	}
 </script>
 
 <Navbar {fluid} color="default">
@@ -53,8 +39,6 @@
 	</div>
 
 	<div class="ms-auto flex items-center gap-2 p-1">
-		<p class="text-sm dark:text-gray-300">Language:</p>
-		<Select class="mt-2" items={locales} bind:value={selected} placeholder="Select language" />
 		<a
 			class="github-button"
 			href="https://github.com/thymis-io/thymis"
@@ -66,5 +50,6 @@
 		</a>
 		<UserMenu />
 		<DarkMode />
+		<LanguageSelect />
 	</div>
 </Navbar>
